+++ src/Application/Command/Event/Update.php
@@ -44,6 +44,9 @@ class Update extends AbstractEvent
     /** @var bool */
     public $linkedinLoginEnabled;
 
+    /** @var bool */
+    public $accessControlEnabled;
+
     /**
      * @param Model\Event $event
      */
@@ -73,6 +76,7 @@ class Update extends AbstractEvent
         $this->visio = $event->getConfiguration()->isVisio();
         $this->googleLoginEnabled = $event->isGoogleLoginEnabled();
         $this->linkedinLoginEnabled = $event->isLinkedinLoginEnabled();
+        $this->accessControlEnabled = $event->isAccessControlEnabled();
 
         foreach ($event->getTranslations() as $translation) {
             $this->translations[$translation->getLocale()] = [