+++ src/Domain/Model/Event.php
@@ -132,6 +132,9 @@ class Event implements EventInterface, TraceableInterface
     /** @var bool */
     private $linkedinLoginEnabled;
 
+    /** @var bool */
+    private $accessControlEnabled;
+
     public function __construct(
         string $title,
         string $fallback,
@@ -151,7 +154,8 @@ class Event implements EventInterface, TraceableInterface
         bool $disabledEmailChanging = false,
         bool $disabledPasswordChanging = false,
         bool $googleLoginEnabled = false,
-        bool $linkedinLoginEnabled = false
+        bool $linkedinLoginEnabled = false,
+        bool $accessControlEnabled = false
     ) {
         $this->translations   = new ArrayCollection();
         $this->configuration  = new Configuration();
@@ -179,6 +183,7 @@ class Event implements EventInterface, TraceableInterface
         $this->disabledPasswordChanging = $disabledPasswordChanging;
         $this->googleLoginEnabled = $googleLoginEnabled;
         $this->linkedinLoginEnabled = $linkedinLoginEnabled;
+        $this->accessControlEnabled = $accessControlEnabled;
     }
 
     /**
@@ -454,6 +459,7 @@ class Event implements EventInterface, TraceableInterface
      * @param bool        $disabledPasswordChanging
      * @param bool        $googleLoginEnabled
      * @param bool        $linkedinLoginEnabled
+     * @param bool        $accessControlEnabled
      */
     public function update(
         $title,
@@ -473,7 +479,8 @@ class Event implements EventInterface, TraceableInterface
         bool $disabledEmailChanging,
         bool $disabledPasswordChanging,
         bool $googleLoginEnabled = false,
-        bool $linkedinLoginEnabled = false
+        bool $linkedinLoginEnabled = false,
+        bool $accessControlEnabled = false
     ) {
         $this->title = $title;
         $this->locales = $locales;
@@ -493,6 +500,7 @@ class Event implements EventInterface, TraceableInterface
         $this->disabledPasswordChanging = $disabledPasswordChanging;
         $this->googleLoginEnabled = $googleLoginEnabled;
         $this->linkedinLoginEnabled = $linkedinLoginEnabled;
+        $this->accessControlEnabled = $accessControlEnabled;
     }
 
     /**
@@ -845,6 +853,11 @@ class Event implements EventInterface, TraceableInterface
     public function hasOAuth2LoginEnabled(): bool
     {
         return $this->isGoogleLoginEnabled() || $this->isLinkedinLoginEnabled();
+    }
+
+    public function isAccessControlEnabled(): bool
+    {
+        return $this->accessControlEnabled;
     }
 
     /**